pyproject.toml
[tool.poetry]
name = "meshroom"
version = "0.2.0"
description = "a commandline Cybersecurity Mesh Assistant"
authors = ["Jerome Fellus <jerome.fellus@sekoia.io>"]
packages = [{include = "meshroom"}]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
websockets = "^13.0.1"
click = "^8.1.7"
pyyaml = "^6.0.2"
tabulate = "^0.9.0"
gitpython = "^3.1.43"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
websocket-client = "^1.8.0"
pydantic = "^2.10.4"
passpy = "^1.0.2"


[tool.poetry.scripts]
meshroom = "meshroom.cli.cli:meshroom"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 190

[tool.pytest.ini_options]
testpaths = ["./tests"]
norecursedirs = ["./tests/e2e/data", "data"]